examples/model_selection/Trails/internal/ml/model_selection/requirement.txt
aiofiles==23.1.0
blessed==1.20.0
certifi==2023.7.22
charset-normalizer==3.2.0
ConfigSpace==0.7.1
contourpy==1.1.0
cycler==0.11.0
einops==0.7.0
fonttools==4.41.0
fvcore==0.1.5.post20221221
gpustat==1.1
h5py==3.10.0
html5tagger==1.3.0
httptools==0.6.0
idna==3.4
importlib-resources==6.0.0
iopath==0.1.10
joblib==1.3.1
kiwisolver==1.4.4
matplotlib==3.7.2
more-itertools==9.1.0
multidict==6.0.4
numpy==1.24.4
nvidia-ml-py==12.535.77
objgraph==3.6.0
orjson==3.9.2
packaging==23.1
palettable==3.3.3
pandas==2.0.3
Pillow==10.0.0
portalocker==2.8.2
psutil==5.9.5
psycopg2-binary==2.9.6
Pympler==1.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
sanic==23.6.0
sanic-routing==23.6.0
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sklearn==0.0
tabulate==0.9.0
termcolor==2.3.0
thop @ git+https://github.com/Lyken17/pytorch-OpCounter.git@43c064afb71383501e41eaef9e8c8407265cf77f
threadpoolctl==3.1.0
torch==1.8.1
torchaudio==0.8.1
torchinfo==1.8.0
torchvision==0.9.1
tqdm==4.47.0
tracerite==1.1.0
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.8.0
urllib3==2.0.4
uvloop==0.17.0
wcwidth==0.2.6
websockets==11.0.3
yacs==0.1.8
zipp==3.16.2